requirements.txt
# Phishing Detection System - Dependencies
# Python 3.8+ recommended

# Core ML & Data
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Web & API
flask>=3.0.0
requests>=2.31.0

# Optional: Domain/WHOIS features (graceful fallback if unavailable)
python-whois>=0.8.0
dnspython>=2.4.0

# Development (optional)
# pytest>=7.0.0
# black>=23.0.0
